// There seems to be a way to spawn particle.
It has to be done, by editing the GSD snapshot's particle parameters.
- Adding to the total number of particles
- Setting all the parameters for every new entity.
- Then re-initializing the calculations.

I am unsure how much overhead does this mean. We seem to lose the analogue nature of nature this way.
Might not even be feasible to use the package, for large number of cells.
This is definitely something that needs to be addressed.