// For clarity, this is the way Lívia suggested adding a new particle to the system (snapshot)
snapshot.particle.resize(oldLength+1)
hoomdSystemObject.restore_snapshot(snapshot)

This way we seem to generate a new particle, where all the properties are default. It seems not neccesary default by hoomd, but rather the unerlying engine, like numpy. So the position and typeid seems to be set to '0' / [0,0,0]
Even with this version I will have to set all the parameters by hand, but still saves:
* Saving a snapshot to disk
* Altering the snapshot
* Reloading the snapshot
* Re-intializing the system

Now we only seem to do:
* hoomdSystemObject.restore_snapshot(snapshot) - Reloading the snapshot
* I think it is "only" reacreating the arrays in memory, I will run benchmarks.